// features/instagram/instagram.types.ts
/**
 * --------------------------------------------------------------------------
 * File:
 * features/instagram/instagram.types.ts
 *
 * Description:
 * Instagram data type definitions.
 *
 * --------------------------------------------------------------------------
 */


export type InstagramPost = {

  id: string;

  image: string;

  alt: string;

  href?: string;

};